Include HMR refresh hash in `"use cache"` cache keys
Today, when editing server components that use `"use cache"` functions,
only updates to the uncached parts of the component will be reflected
in the rendered result after the HMR refresh was applied. Any cached
functions/components will show the same content as before the edit.

This is problematic when a function with a `"use cache"` directive is
edited. In this case, the edits won't be applied, and stale content is
shown in the browser, until the dev server is restarted.

With this PR, we are now including an HMR refresh hash in the cache key
for all `"use cache"` functions. This ensures that those functions
are revalidated when a server component is edited, which avoids showing
stale content.

For Webpack, a real content hash is used, which means that if an edit is
reverted again, the previously cached data will be retrieved from the
cache. For Turbopack, the HRM hash is currently just an incrementing
counter, which may be improved in a follow-up PR.